[DISCUSS BEFORE UPSTREAM] klayout lvs: remove the _dn diodes
This is because these diodes do not have spice model associated and are only
supported within klayout (no magic, xschem, ngspice, ...). Rather than
add devices with uncertain models in ngspice, xschem, ... it makes more
sense to remove them. It could be nice to have them again in the future
if there is some work to get a proper simulation model.