drc: improve topcell logic
Before, multiple things could lead to a silent error that would lead to
a wrongly "DRC clean" statement by klayout. This could be caused by:
* Having multiple topcells and no topcell set
* Having an invalid topcell set

Additionnaly, if no topcell name is provided through the gui options,
the cell that gets verified will be the currently displayed cell. This
changes from the previous default, which was the top cell (which worked
only if there was a single one).
It is also possible now to set a cell name that is not a top cell.