lvs: do not implicitely connect nets that have the same name
This is because it can create issues when running LVS in flat mode.
If you have net1 and net2 named as follows:
net1: TOP:A, child1:A
net2: TOP:N/A, child2:A
Klayout will produce, in flat mode, a net name of A for both nets, even
though they are completely unrelated.