[DISCUSS ME] Use isosub for substrate connection
Adds the ISOSUB capability.
Before, there was only one substrate for any design, and by design it
was not possible to connect to it through contacts, only device
extraction.
This prevented designs to pass LVS, such as basic ones like:
https://github.com/iic-jku/IIC-OSIC-TOOLS/issues/211
as they short the substrate to VSS.

More complicated designs like the foundry pads require on top of that
multiple substrate power domain, such as VSS DVSS separation.

This commit add substrate domains by using polygon to connect the
different zones. For isosub, it is easy: simply use the existing ones.
For the default substrate, it is necessary to make a negative (chip -
isosub). For getting the chip area, I favor the border layer, and added
a fallback to the bounding box of the design. In order to make sure that
two default substrate area are going to be connected even if they are
separated by an isosub polygon that spans the entire bounding box, I
upsize the chip area by one database unit.
The other problem is that touching polygons are considered as shorted by
klayout. That means that I had to upsize the isosub polygons during the
subtraction.
This means that 2 connection to the substrate outside of isosub polygons
connected by a thin strip of database unit width will not be considered
as connected in LVS.